backend/requirements.txt
fastapi
uvicorn
boto3
pydantic_settings
dotenv
sqlalchemy
psycopg2-binary